# frontend/src/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AcquireLeaseRequest": {
        "properties": {
          "session_id": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,255}$",
            "title": "Session Id",
            "type": "string"
          },
          "symbol": {
            "maxLength": 32,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9^][A-Za-z0-9._:^=-]{0,31}$",
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "symbol"
        ],
        "title": "AcquireLeaseRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeaseTokenRequest": {
        "properties": {
          "lease_token": {
            "maxLength": 256,
            "minLength": 20,
            "title": "Lease Token",
            "type": "string"
          }
        },
        "required": [
          "lease_token"
        ],
        "title": "LeaseTokenRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "可信行情、AI Prediction 与证据化研究 API",
    "title": "FinSight API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/monitor/comments": {
      "get": {
        "operationId": "list_monitor_comments_api_monitor_comments_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,255}$",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9^][A-Za-z0-9._:^=-]{0,31}$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Monitor Comments",
        "tags": [
          "Monitor"
        ]
      }
    },
    "/api/monitor/comments/stream": {
      "get": {
        "operationId": "stream_monitor_comments_api_monitor_comments_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,255}$",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9^][A-Za-z0-9._:^=-]{0,31}$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "last_event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Monitor Comments",
        "tags": [
          "Monitor"
        ]
      }
    },
    "/api/monitor/leases": {
      "post": {
        "operationId": "acquire_monitor_lease_api_monitor_leases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcquireLeaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acquire Monitor Lease",
        "tags": [
          "Monitor"
        ]
      }
    },
    "/api/monitor/leases/{lease_id}": {
      "delete": {
        "operationId": "release_monitor_lease_api_monitor_leases__lease_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "lease_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lease Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Monitor Lease",
        "tags": [
          "Monitor"
        ]
      },
      "put": {
        "operationId": "renew_monitor_lease_api_monitor_leases__lease_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "lease_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Lease Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Renew Monitor Lease",
        "tags": [
          "Monitor"
        ]
      }
    },
...
  }
}